refactor: Dto request/response 구분
