+++ src/main/java/com/st/eighteen_be/user/api/UserApiController.java
@@ -4,8 +4,8 @@ import com.st.eighteen_be.common.response.ApiResp;
 import com.st.eighteen_be.common.security.SecurityUtil;
 import com.st.eighteen_be.jwt.JwtTokenDto;
 import com.st.eighteen_be.user.domain.UserInfo;
-import com.st.eighteen_be.user.dto.sign.SignInRequestDto;
-import com.st.eighteen_be.user.dto.sign.SignUpRequestDto;
+import com.st.eighteen_be.user.dto.request.SignInRequestDto;
+import com.st.eighteen_be.user.dto.request.SignUpRequestDto;
 import com.st.eighteen_be.user.service.UserService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
