+++ src/test/java/com/st/eighteen_be/user/controller/UserApiControllerTest.java
@@ -6,8 +6,8 @@ import com.st.eighteen_be.jwt.JwtTokenDto;
 import com.st.eighteen_be.jwt.JwtTokenProvider;
 import com.st.eighteen_be.user.WithCustomMockUser;
 import com.st.eighteen_be.user.domain.UserInfo;
-import com.st.eighteen_be.user.dto.sign.SignInRequestDto;
-import com.st.eighteen_be.user.dto.sign.SignUpRequestDto;
+import com.st.eighteen_be.user.dto.request.SignInRequestDto;
+import com.st.eighteen_be.user.dto.request.SignUpRequestDto;
 import com.st.eighteen_be.user.repository.UserRepository;
 import com.st.eighteen_be.user.service.UserService;
 import org.junit.jupiter.api.BeforeEach;
