+++ src/test/java/com/st/eighteen_be/user/service/UserServiceTest.java
@@ -7,8 +7,8 @@ import com.st.eighteen_be.jwt.JwtTokenProvider;
 import com.st.eighteen_be.token.domain.RefreshToken;
 import com.st.eighteen_be.token.repository.RefreshTokenRepository;
 import com.st.eighteen_be.user.domain.UserInfo;
-import com.st.eighteen_be.user.dto.sign.SignInRequestDto;
-import com.st.eighteen_be.user.dto.sign.SignUpRequestDto;
+import com.st.eighteen_be.user.dto.request.SignInRequestDto;
+import com.st.eighteen_be.user.dto.request.SignUpRequestDto;
 import com.st.eighteen_be.user.repository.UserRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
